Replacing router logic in routes.createRouter
