+++ lib/components/DevCards.js
@@ -1,5 +1,4 @@
 var createHistory = require('history').createHashHistory;
-var router = require('../router');
 
 var React = require('react');
 
@@ -9,6 +8,7 @@ var utils = require('../utils');
 var find = utils.find;
 
 var routes = require('../routes');
+var createRouter = routes.createRouter;
 var createHref = routes.createHref;
 
 var DevCardNamespace = require('./DevCardNamespace');
@@ -45,22 +45,11 @@ var DevCards = React.createClass({
 
     var setState = this.setState.bind(this);
 
-    // Listening to URL changes
-    this.release = history.listen(function(location) {
-      var match = router.match(location.pathname);
+    var notFound = function() {
+      history.push('/');
+    };
 
-      // If we don't match, we redirect to /
-      if (!match) {
-        return history.push('/');
-      }
-
-      var params = match.params;
-
-      setState({
-        namespace: params.namespace || null,
-        card: params.card || null
-      });
-    });
+    this.release = history.listen(createRouter(setState, notFound));
   },
   componentWillUnmount: function() {
     this.release();
